General changes

+++ src/main/java/net/planetgeeks/mcstarter/util/task/ProgressTask.java
@@ -3,6 +3,7 @@ package net.planetgeeks.mcstarter.util.task;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
+import net.planetgeeks.mcstarter.util.ProgressView;
 
 /**
  * An extended {@link #Task} whose progress can be monitored.
@@ -16,7 +17,7 @@ public abstract class ProgressTask<T> extends Task<T> implements ProgressView
 	private double progress;
 	
 	@Override
-    public double getProgress(int scale)
+    public double getProgress(double scale)
     {
     	return getProgress() * scale;
     }
